pyproject.toml
[tool.poetry]
name = "ubimaior"
version = "0.1.0"
description = "Manage hierarchy of objects as if they were one"
license = "Apache-2.0 OR MIT"
authors = ["Massimiliano Culpo <massimiliano.culpo@gmail.com>"]
repository = "https://github.com/forthelols/ubimaior.git"
homepage = "https://github.com/forthelols/ubimaior"
documentation = "https://ubimaior.readthedocs.io"
readme = "README.rst"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Programming Language :: Python :: 2",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = "^3.6"
jsonschema = "^3.2.0"
click = "^7.1.2"
"ruamel.yaml" = {version = "^0.16.12", optional = true}
toml = {version = "^0.10.1", optional = true }

[tool.poetry.extras]
YAML = ["ruamel.yaml"]
TOML = ["toml"]

[tool.poetry.dev-dependencies]
pytest = ">=5.3.2"
pytest-cov = ">=2.8.1"
tox = "^3.20.1"
pylint = "^2.7.0"
flake8 = "^3.8.4"
coverage = "^5.4"
vermin = "^1.1.0"
black = "^20.8b1"

[tool.pylint."MESSAGES CONTROL"]
disable = "fixme,super-with-arguments,raise-missing-from"

[tool.black]
line-length = 100
target-version = ['py27', 'py37']

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
